backend/requirements.txt
flask
flask-cors
firebase-admin
google-cloud-vision
python-dotenv
werkzeug
